chore(nexus): core module for core traits
