+++ bins/todo/Cargo.toml
@@ -6,6 +6,7 @@ license.workspace = true
 authors.workspace = true
 
 [dependencies]
+async-trait = { workspace = true }
 nexus = { path = "../../crates/nexus" }
 rusqlite = { workspace = true, features = ["bundled"] }
 workspace-hack = { version = "0.1.0", path = "../../crates/workspace-hack" }
